fix: move Score below Email in Grade struct
Go's json library will marshal structs in the order that they are
defined. In this case, I want the order "email, score" because that
order reads better for people entering grades in gradebook files.

I have to add the //nolint directive because otherwise govet will try to
optimize the struct. In this case, I can live with the extra bytes.